+++ src/components/forms/TextFieldList/TextFieldList.tsx
@@ -46,7 +46,16 @@ export const TextFieldList: React.FC<TextFieldListProps> = props => {
   useEffect(() => onValueChange(values.map(value => value.text)), [values])
   useEffect(() => onErrors(flatErrorMessages), [errorMessages])
   useEffect(() => {
-    setValues(intialValues)
+    const mergedValue = props.value.reduce((aggr, curr, index) => {
+      const { id = RandomString.generate(20) } = values[index]
+
+      return [...aggr, {
+        id: id,
+        text: curr
+      }]
+    }, [])
+
+    setValues(mergedValue)
   }, [...(props.value || [])])
 
   const {