+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project **doesn't** adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) for now.
 
 ## [Unreleased]
+### Changed
+- We wrapp the `DatePickerField` with `UiKitInitializer` only on its storybook to display
+  the documentation with translation support.
+
 ### Fixed
 - Correct translated strings are display by `DatePickerField`
+- Change how `UiKitInitializer` manage to use the i18n delegate (`this` problem, null pointer exception problem)
 
 ## [v0.0.27] - 2019-02-11
 ### Added