+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project **doesn't** adhere to [Semantic Versioning](http://semver.org/s
 ### Changed
 - We wrapp the `DatePickerField` with `UiKitInitializer` only on its storybook to display
   the documentation with translation support.
+- The `RadioButton` can be disabled even on a non-disabled `RadioGroup`.
 
 ### Fixed
 - Correct translated strings are display by `DatePickerField`