+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) since [v0.1.0].
 
 ## [Unreleased]
+### Fixed
+- `TextFieldList` do not loose focus when the user is writing something.
 
 ## [v0.1.0] - 2019-02-13
 ### Added